Move presence updating loop to the main bot class
Since this has to do with the rich presence of the bot, it's much better suited for the bot class as opposed to the `Util` module
